+++ cmd/zombiezen-sqlite-migrate/testdata/skeleton/crawshaw.io/sqlite/sqlitex/stubs.go
@@ -1,4 +1,4 @@
-// Copyright 2021 Ross Light
+// Copyright 2021 Roxy Light
 // SPDX-License-Identifier: ISC
 
 // Test stubs for crawshaw.io/sqlite/sqlitex.