+++ sqlite.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2018 David Crawshaw <david@zentus.com>
-// Copyright (c) 2021 Ross Light <ross@zombiezen.com>
+// Copyright (c) 2021 Roxy Light <roxy@zombiezen.com>
 //
 // Permission to use, copy, modify, and distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -65,6 +65,15 @@ const ptrSize = types.Size_t(unsafe.Sizeof(uintptr(0)))
 //   - [OpenWAL]
 //   - [OpenURI]
 //
+// The OpenWAL flag is an extension specific to this library:
+// it runs "PRAGMA journal_mode=wal;" before returning.
+// While convenient, this constitutes a write transaction
+// and may fail if the database is contended.
+// A default timeout of a few seconds is used
+// so that OpenConn does not block indefinitely.
+// If you want to have more control over the timeout on setting WAL,
+// do not pass OpenWAL and execute "PRAGMA journal_mode=wal;" yourself.
+//
 // https://www.sqlite.org/c3ref/open.html
 func OpenConn(path string, flags ...OpenFlags) (*Conn, error) {
 	var openFlags OpenFlags
@@ -115,7 +124,6 @@ func OpenConn(path string, flags ...OpenFlags) (*Conn, error) {
 	if openFlags&OpenWAL != 0 {
 		// Set timeout for enabling WAL.
 		// See https://github.com/crawshaw/sqlite/pull/113 for details.
-		// TODO(maybe): Pass in Context to OpenConn?
 		c.SetBusyTimeout(10 * time.Second)
 
 		stmt, _, err := c.PrepareTransient("PRAGMA journal_mode=wal;")
@@ -175,7 +183,7 @@ func openConn(path string, openFlags OpenFlags) (_ *Conn, err error) {
 	}
 	if c.conn == 0 {
 		// Not enough memory to allocate the sqlite3 object.
-		return nil, fmt.Errorf("sqlite: open %q: %w", path, sqliteError{res})
+		return nil, fmt.Errorf("sqlite: open %q: %w", path, res.ToError())
 	}
 	if res != ResultOK {
 		// sqlite3_open_v2 may still return a sqlite3* just so we can extract the error.
@@ -184,7 +192,7 @@ func openConn(path string, openFlags OpenFlags) (_ *Conn, err error) {
 			res = extres
 		}
 		lib.Xsqlite3_close_v2(tls, c.conn)
-		return nil, fmt.Errorf("sqlite: open %q: %w", path, sqliteError{res})
+		return nil, fmt.Errorf("sqlite: open %q: %w", path, res.ToError())
 	}
 	lib.Xsqlite3_extended_result_codes(tls, c.conn, 1)
 
@@ -204,6 +212,9 @@ func (c *Conn) Close() error {
 	if c == nil {
 		return fmt.Errorf("sqlite: close: nil connection")
 	}
+	if c.closed {
+		return fmt.Errorf("sqlite: close: already closed")
+	}
 	c.cancelInterrupt()
 	c.closed = true
 	for _, stmt := range c.stmts {
@@ -220,7 +231,7 @@ func (c *Conn) Close() error {
 	delete(allConns.table, c.conn)
 	allConns.mu.Unlock()
 	if !res.IsSuccess() {
-		return fmt.Errorf("sqlite: close: %w", sqliteError{res})
+		return fmt.Errorf("sqlite: close: %w", res.ToError())
 	}
 	return nil
 }
@@ -381,7 +392,7 @@ func busyHandlerCallback(tls *libc.TLS, pArg uintptr, count int32) int32 {
 func (c *Conn) interrupted() error {
 	select {
 	case <-c.doneCh:
-		return sqliteError{ResultInterrupt}
+		return ResultInterrupt.ToError()
 	default:
 		return nil
 	}
@@ -467,7 +478,7 @@ func (c *Conn) Prepare(query string) (*Stmt, error) {
 // https://www.sqlite.org/c3ref/prepare.html
 func (c *Conn) PrepareTransient(query string) (stmt *Stmt, trailingBytes int, err error) {
 	if c == nil {
-		return nil, 0, fmt.Errorf("sqlite: prepare %q: nil connection", query)
+		return nil, 0, fmt.Errorf("sqlite: prepare: nil connection")
 	}
 	// TODO(soon)
 	// if stmt != nil {
@@ -477,7 +488,11 @@ func (c *Conn) PrepareTransient(query string) (stmt *Stmt, trailingBytes int, er
 	// 		}
 	// 	})
 	// }
-	return c.prepare(query, 0)
+	stmt, trailingBytes, err = c.prepare(query, 0)
+	if err != nil {
+		err = fmt.Errorf("sqlite: prepare: %w", err)
+	}
+	return stmt, trailingBytes, err
 }
 
 func (c *Conn) prepare(query string, flags uint32) (*Stmt, int, error) {
@@ -502,6 +517,10 @@ func (c *Conn) prepare(query string, flags uint32) (*Stmt, int, error) {
 	defer libc.Xfree(c.tls, stmtPtr)
 	res := ResultCode(lib.Xsqlite3_prepare_v3(c.tls, c.conn, cquery, -1, flags, stmtPtr, ctrailingPtr))
 	if err := c.extreserr(res); err != nil {
+		if offset, ok := ErrorOffset(err); ok && offset <= len(query) {
+			line, col := linecol(query[:offset])
+			err = fmt.Errorf("%d:%d: %w", line, col, err)
+		}
 		return nil, 0, err
 	}
 	ctrailing := *(*uintptr)(unsafe.Pointer(ctrailingPtr))
@@ -619,10 +638,33 @@ func (c *Conn) extreserr(res ResultCode) error {
 	if res.IsSuccess() {
 		return nil
 	}
-	if msg := libc.GoString(lib.Xsqlite3_errmsg(c.tls, c.conn)); msg != "" {
-		return fmt.Errorf("%w: %s", res.ToError(), msg)
+	return &extendedError{
+		code:    res,
+		offset:  int(lib.Xsqlite3_error_offset(c.tls, c.conn)),
+		message: libc.GoString(lib.Xsqlite3_errmsg(c.tls, c.conn)),
 	}
-	return res.ToError()
+}
+
+// linecol computes the 1-based line number and column
+// of a character placed after s.
+func linecol(s string) (line, col int) {
+	line, col = 1, 1
+	for _, c := range s {
+		// This is intentionally a bit naive with regards to control characters.
+		// If the user wants greater control,
+		switch c {
+		case '\n':
+			line++
+			col = 1
+		case '\t':
+			const tabWidth = 8
+			locationInTab := (col - 1) % tabWidth
+			col += tabWidth - locationInTab
+		default:
+			col++
+		}
+	}
+	return
 }
 
 // Stmt is an SQLite3 prepared statement.
@@ -645,7 +687,7 @@ type Stmt struct {
 
 func (stmt *Stmt) interrupted() error {
 	if stmt.prepInterrupt {
-		return sqliteError{ResultInterrupt}
+		return ResultInterrupt.ToError()
 	}
 	return stmt.conn.interrupted()
 }